Caches the gcc-arm.zip file for the ports_windows workflow.
This is an initial test of Github caching. (Partial #2969)
Ref: https://docs.github.com/en/actions/guides/caching-dependencies-to-speed-up-workflows

There are issues with caching on Windows runners, so the implementation I've
come up with does the cache creation and downloading on a Linux runner and
only accesses it from Windows.

For Windows we have to use GNU tar instead BSD tar
Ref: https://github.com/actions/cache/issues/576
Otherwise we don't get support for .zsdt compression which the cache requires.